Better event listener strategy
Change the event listener to fire for controller beforeRedirect
events and then redirect to the CoInvites controller with the
verifyEmailAddress action and pass the ID of the email just
edited. That will display the usual flash for the user
and kick off the email verification flow. The final redirect
is back to the CoPerson canvas.