Allow platform or CO admin to edit verification status
Allow a platform or CO admin to edit the email verification
status. The strategy is to add to the existing CakeEvent
listener logic to trap an edit and if the actor is a
CO admin or platform admin then redirect into the plugin
which prevents a different view where the verifification status
can be edited. The edit action for the plugin is only avaiable
to CO or platform admins.